classification_krawtchouk_3d/requirements.txt
numpy
scipy
matplotlib
pandas
scikit-learn
scikit-image
pyyaml
tqdm
torch
